按照RBAC重新设计数据库，加入spring security

+++ src/main/java/com/lushu/checksystem/pojo/File.java
@@ -9,9 +9,9 @@ import java.util.Date;
 @Data
 public class File implements Serializable {
 
-  private BigInteger id;
+  private Integer id;
   private String name;
-  private Float size;
+  private BigInteger size;
   private Date updateTime;
   private Integer type;
   private String permission;
