+++ src/main/java/com/lushu/checksystem/pojo/Inform.java
@@ -3,14 +3,13 @@ package com.lushu.checksystem.pojo;
 import lombok.Data;
 
 import java.io.Serializable;
-import java.math.BigInteger;
 
 @Data
 public class Inform implements Serializable {
 
-  private BigInteger id;
-  private BigInteger sendId;
-  private BigInteger receiveId;
+  private Integer id;
+  private Integer sendId;
+  private Integer receiveId;
   private String content;
